Fixed some imports and organized some state functions into accurate files.
