+++ ui/state/index.ts
@@ -1,3 +1,5 @@
 export * from "./campaigns"
+export * from "./client"
+export * from "./dao"
 export * from "./global"
-export * from "./web3"
+export * from "./token"
